+++ document-downloader.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name:  Document Downloader
  * Description:  Document Download Manager
- * Version:      1.0.9(beta)
+ * Version:      1.1.0(beta)
  * Requires PHP: 7.4
  * Author:       Alan Blair<alan@wpeasy.au>
  * Text Domain:  document-downloader